+++ g_monster.c
@@ -17,7 +17,7 @@ void monster_fire_bullet (edict_t *self, vec3_t start, vec3_t dir, int damage, i
         fire_bullet (self, start, dir, damage, kick, hspread, vspread, MOD_UNKNOWN);
 
         gi.WriteByte (svc_muzzleflash2);
-        gi.WriteShort (self - g_edicts);
+        gi.WriteShort ((int)(self - g_edicts));
         gi.WriteByte (flashtype);
         gi.multicast (start, MULTICAST_PVS);
 }
@@ -27,7 +27,7 @@ void monster_fire_shotgun (edict_t *self, vec3_t start, vec3_t aimdir, int damag
         fire_shotgun (self, start, aimdir, damage, kick, hspread, vspread, count, MOD_UNKNOWN);
 
         gi.WriteByte (svc_muzzleflash2);
-        gi.WriteShort (self - g_edicts);
+        gi.WriteShort ((int)(self - g_edicts));
         gi.WriteByte (flashtype);
         gi.multicast (start, MULTICAST_PVS);
 }
@@ -37,7 +37,7 @@ void monster_fire_blaster (edict_t *self, vec3_t start, vec3_t dir, int damage, 
         fire_blaster (self, start, dir, damage, speed, effect, false);
 
         gi.WriteByte (svc_muzzleflash2);
-        gi.WriteShort (self - g_edicts);
+        gi.WriteShort ((int)(self - g_edicts));
         gi.WriteByte (flashtype);
         gi.multicast (start, MULTICAST_PVS);
 }       
@@ -47,7 +47,7 @@ void monster_fire_grenade (edict_t *self, vec3_t start, vec3_t aimdir, int damag
         fire_grenade (self, start, aimdir, damage, speed, 2.5, damage+40);
 
         gi.WriteByte (svc_muzzleflash2);
-        gi.WriteShort (self - g_edicts);
+        gi.WriteShort ((int)(self - g_edicts));
         gi.WriteByte (flashtype);
         gi.multicast (start, MULTICAST_PVS);
 }
@@ -57,7 +57,7 @@ void monster_fire_rocket (edict_t *self, vec3_t start, vec3_t dir, int damage, i
         fire_rocket (self, start, dir, damage, speed, damage+20, damage);
 
         gi.WriteByte (svc_muzzleflash2);
-        gi.WriteShort (self - g_edicts);
+        gi.WriteShort ((int)(self - g_edicts));
         gi.WriteByte (flashtype);
         gi.multicast (start, MULTICAST_PVS);
 }       
@@ -67,7 +67,7 @@ void monster_fire_railgun (edict_t *self, vec3_t start, vec3_t aimdir, int damag
         fire_rail (self, start, aimdir, damage, kick);
 
         gi.WriteByte (svc_muzzleflash2);
-        gi.WriteShort (self - g_edicts);
+        gi.WriteShort ((int)(self - g_edicts));
         gi.WriteByte (flashtype);
         gi.multicast (start, MULTICAST_PVS);
 }
@@ -77,7 +77,7 @@ void monster_fire_bfg (edict_t *self, vec3_t start, vec3_t aimdir, int damage, i
         fire_bfg (self, start, aimdir, damage, speed, damage_radius);
 
         gi.WriteByte (svc_muzzleflash2);
-        gi.WriteShort (self - g_edicts);
+        gi.WriteShort ((int)(self - g_edicts));
         gi.WriteByte (flashtype);
         gi.multicast (start, MULTICAST_PVS);
 }
